Database rewrited (once again) Re-writed for beta-09 from SurrealDB, that doesn't accept Datastore and session anymore.
