+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 [dependencies]
 actix-web = "4.2.1"
 anyhow = "1.0.68"
-surrealdb = "1.0.0-beta.8"
+surrealdb ={ git = "https://github.com/surrealdb/surrealdb" }
 tokio = { version = "1.24.2", features = ["full"] }
 rust-argon2 = "1.0"
 serde = { version = "1.0.130", features = ["derive"] }
