+++ src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod hash_password;
+pub mod strings;
